feat(#17): layout page joined, join 구별
layout page에 참가한 페이지, 참가못한 페이지 구별
