+++ libs/ui/src/index.ts
@@ -6,3 +6,5 @@ export * from './lib/sidebar';
 export * from './lib/serverModal';
 export * from './lib/LayoutPage';
 export * from './context/ServerCreateProvider';
+export * from './context/ServerJoinedTeamProvider'
+export * from './lib/TeamJoined';
